[typer] debug
Significant issues:
Ne(VDef) does not expand in case of no applications, such as Bool;
In HOAS closure of the value of Lam and Pi, wrongly pass value environments before binding variables, it had better create a closure create API to avoid such mistakes.
In our implementation, Def is not in the local enviroments, in this case, eval_app_bds should allow value environment to be longer than binds list.